move `TestStore` to main tests file
